+++ packages/radonis-server/src/types/index.ts
@@ -23,5 +23,6 @@ export interface DocumentProps {
 export interface HydrationRootProps {
   children: ReactElement<Record<string, any>>
   component: ComponentIdentifier
+  className?: string
   disabled?: boolean
 }