+++ CHANGELOG.md
@@ -10,9 +10,11 @@ All notable changes to this project will be documented in this file.
 - BREAKING: Renamed `afterOutput` plugin hook to `afterOutputAssets`.
 - BREAKING: Increased minimum supported ECMAScript version to `ES2022` (NodeJS 16 and up).
 - Removed and deprecated internal package `@microeinhundert/radonis-manifest`.
+- Radonis packages now ship with source maps.
+- HydrationRoots now accept the `className` prop.
+- Nested HydrationRoots are now ignored instead of throwing an error.
 - Improved exceptions.
 - Internal refactorings related to dependency injection.
-- Ship radonis packages with source maps.
 
 ## [1.10.1] - 2022-09-28